+++ src/temp/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>xPro</name>
+	<name>xpro</name>
 	<comment></comment>
 	<projects>
 	</projects>
